requirements.txt
keras==2.9.0
matplotlib==3.5.2
numpy==1.23.0
pandas==1.4.3
python_dateutil==2.8.2
scikit_learn==1.1.1
streamlit==1.10.0
tensorflow==2.9.1
yfinance==0.1.72
